🚧 normalize_cubdata関数をファイル分けしたり、tokenize_linesを一旦作ったり
