+++ map/test/simple.cub
@@ -1,0 +1,11 @@
+NO ./path_to_the_north_texture
+SO ./path_to_the_south_texture
+WE ./path_to_the_west_texture
+EA ./path_to_the_east_texture
+F 220,100,0
+C 225,30,0
+111111
+100101
+101001
+1100N1
+111111
